fix(history): Fix invalid database access on failed schema upgrade
dbSchemaUpgrade attempts to reset the db pointer when failures occur,
however this only resets the shadowed db pointer that has been copied
into the member variable. This allows isValid() calls to pass when they
shouldn't.

This fix prevents the shadowing of the member db variable ensuring that
isValid() fails on bad databases.